refactor(examples): migrate soi_waveguide + step_index_fiber + fiber_smf28_solve_timing to standalone crates
Move the waveguide/fiber-group examples out of `crates/geode-core/examples`
into top-level standalone crates built on the shared `geode-app` harness
(Epic #398 Phase 3b), matching the mie_sphere pilot and the Phase 3a
driven-scattering/open-quasimode migrations.

For each example:
- `git mv` the body into `examples/<name>/src/main.rs`.
- Add an `examples/<name>/Cargo.toml` with workspace-inherited deps
  (clap, geode-app, geode-core) — none use faer/burn directly or the
  `#[path] common/viz_export_helper.rs` include, so no extra deps.
- Wrap the body in a `#[derive(clap::Parser)] Args` that flattens
  `geode_app::Verbosity`, impls `geode_app::App`, and routes through
  `fn main() -> ExitCode { geode_app::main::<Args>() }`. These examples
  write fixed `CARGO_MANIFEST_DIR`-relative artifacts, so OutputDir is
  omitted (matching the mie crates).

Physics, report output, and the `results.toml` artifacts are preserved
exactly; the `CARGO_MANIFEST_DIR/../../benchmarks/<name>` walk-up is
unchanged (examples/<name> is the same two levels below the workspace
root as crates/geode-core was). The old `common/` helper and the
remaining un-migrated examples are left untouched for Phase 4.
