+++ examples/fiber_smf28_solve_timing/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "fiber_smf28_solve_timing"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+description = "Performance harness timing the SMF-28 fiber p=2 modal solve on the sparse-direct assembly path (issue #327)."
+
+[[bin]]
+name = "fiber_smf28_solve_timing"
+path = "src/main.rs"
+
+[dependencies]
+clap = { workspace = true }
+geode-app = { workspace = true }
+geode-core = { workspace = true }
+
+# NOTE: this crate deliberately defines NO backend features. The Burn
+# backend (`wgpu` by default, or `ndarray` on headless CI) is selected by
+# geode-core's default features / the workspace `--features` flags at build
+# time, exactly like `geode-cli` and `geode-validation`.
+
+[lints.rust]
+unsafe_code = "forbid"
