+++ examples/soi_waveguide/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "soi_waveguide"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+description = "Silicon-on-insulator (SOI) strip-waveguide benchmark: fundamental quasi-TE n_eff vs the effective-index-method oracle (issue #306, Epic #303 Phase 1C)."
+
+[[bin]]
+name = "soi_waveguide"
+path = "src/main.rs"
+
+[dependencies]
+clap = { workspace = true }
+geode-app = { workspace = true }
+geode-core = { workspace = true }
+
+# NOTE: this crate deliberately defines NO backend features. The Burn
+# backend (`wgpu` by default, or `ndarray` on headless CI) is selected by
+# geode-core's default features / the workspace `--features` flags at build
+# time, exactly like `geode-cli` and `geode-validation`.
+
+[lints.rust]
+unsafe_code = "forbid"
